feat: add bubble sort, selection sort, insertion sort

+++ algorithms/src/main/java/com/examples/fibonacci/FibonacciSeriesUtils.java
@@ -2,14 +2,15 @@ package com.examples.fibonacci;
 
 public class FibonacciSeriesUtils {
 
-    public static int nthFibonacciTermRecursiveMethod(int n) {
+    //The complexity of the algorithm О(Ф^n);
+    public static long nthFibonacciTermRecursiveMethod(int n) {
         if (n == 0 || n == 1) {
             return n;
         }
         return (nthFibonacciTermRecursiveMethod(n - 1) + nthFibonacciTermRecursiveMethod(n - 2));
     }
 
-    private static int nthFibonacciTermCashingMethodUtil(int n, int[] f) {
+    private static long nthFibonacciTermCashingMethodUtil(int n, long[] f) {
 
         if (f[n] == -1) {
             f[n] = (nthFibonacciTermCashingMethodUtil(n - 1, f) + nthFibonacciTermCashingMethodUtil(n - 2, f));
@@ -17,8 +18,9 @@ public class FibonacciSeriesUtils {
         return f[n];
     }
 
-    public static int nthFibonacciTermCashingMethod(int n) {
-        int[] f = new int[n + 1];
+    //The complexity of the algorithm О(N);
+    public static long nthFibonacciTermCashingMethod(int n) {
+        long[] f = new long[n + 1];
         f[0] = 0;
         f[1] = 1;
         for (int i = 2; i <= n; i++) {
@@ -27,12 +29,13 @@ public class FibonacciSeriesUtils {
         return nthFibonacciTermCashingMethodUtil(n, f);
     }
 
-    public static int nthFibonacciTermIterativeMethod(int n) {
+    //The complexity of the algorithm О(N);
+    public static long nthFibonacciTermIterativeMethod(int n) {
         if (n == 0 || n == 1) {
             return n;
         }
-        int n0 = 0, n1 = 1;
-        int tempNthTerm = 0;
+        long n0 = 0, n1 = 1;
+        long tempNthTerm = 0;
         for (int i = 2; i <= n; i++) {
             tempNthTerm = n0 + n1;
             n0 = n1;
@@ -41,9 +44,10 @@ public class FibonacciSeriesUtils {
         return tempNthTerm;
     }
 
-    public static int nthFibonacciTermUsingBinetsFormula(int n) {
+    //The complexity of the algorithm О(1);
+    public static long nthFibonacciTermUsingBinetsFormula(int n) {
         double squareRoutOf5 = Math.sqrt(5);
         double phi = (1 + squareRoutOf5) / 2;
-        return (int) ((Math.pow(phi, n) - Math.pow(-phi, -n)) / squareRoutOf5);
+        return (long) ((Math.pow(phi, n) - Math.pow(-phi, -n)) / squareRoutOf5);
     }
 }
